requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
